requirement.txt
requests
khl.py
python-Levenshtein
fuzzywuzzy